requirements.txt
# Core Dependencies
pystray>=0.19.0
pillow>=10.0.0
keyboard>=0.13.0
pyaudio>=0.2.13
openai>=1.0.0
pyperclip>=1.8.0
pyautogui>=0.9.50
python-dotenv>=1.0.0

# Build Tool
pyinstaller>=6.0.0

# Development & Testing
black>=23.0.0
pylint>=3.0.0
pytest>=7.0.0
watchdog>=3.0.0